hook up progress bar with download events
